Set numeric value for public
With mysql, the previous way was trying to set the value in the db to either 1 or null, requiring the public column to be nullable.   I think this is a mistake now, the column is a boolean, it is either true or false, there shouldn’t be a third option (null).